Documentation clean up work, added examples for CLI helper functions
