+++ example.conf
@@ -1,0 +1,2 @@
+# Comments are ignored
+foo=bar
